feat: Convert taxa/photos page to TSX + responsive styles, WEB-984
PR 3 of the WEB-984 taxon detail page split: the taxa/photos page.

- taxa/photos components app.jsx + photo_browser.jsx -> .tsx (use the
  PR-2 shared TaxonPageHeader / consolidated taxon_crumbs_container)
- delete the per-page taxa/photos/containers/taxon_crumbs_container.js
  (consolidated into taxa/shared in PR 2; only app.jsx imported it)
- photos.scss: responsive PhotoBrowser (container-padding, grid layout,
  mobile controls)
- autocomplete.scss (moved from PR 2): mobile wrap/ellipsis on the
  app-wide taxon-autocomplete widget

Breakpoint alignment with WEB-1026 (carry-forward from PR 2):
- bare max-width queries converted to the .98px -max variants
  ($sm -> $sm-max in photos.scss, $md -> $md-max in autocomplete.scss)
- PhotoBrowser container max-width hardcoded to 1300px (the feature
  branch overloaded $xl=1300 as both breakpoint and container width;
  $xl is now a pure 1200px breakpoint)

typecheck, scoped eslint, and webpack build all pass.
