+++ app/assets/stylesheets/autocomplete.scss
@@ -1,4 +1,5 @@
 @import "./colors";
+@import "./breakpoints";
 
 $img-width: 40px;
 $img-margin: 3px;
@@ -63,6 +64,32 @@ $label-start: 7px;
 }
 .ui-autocomplete.taxon-autocomplete {
   max-width: 500px;
+
+  @media ( max-width: $md-max ) {
+    max-width: calc( 100vw - 30px );
+
+    .ac {
+      min-width: 0;
+    }
+
+    .ac-label {
+      min-width: 0;
+      overflow: hidden;
+
+      > div {
+        min-width: 0;
+        overflow: hidden;
+        width: 100%;
+      }
+    }
+
+    .ac-label .title,
+    .ac-label .subtitle {
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+    }
+  }
 }
 .ac-message[data-type='message'] {
   cursor: default;
