+++ app/assets/stylesheets/taxa/photos.scss
@@ -1,4 +1,9 @@
+@import "../breakpoints";
+
 .PhotoBrowser {
+  @include container-padding;
+  max-width: 1300px;
+  margin: 0 auto;
   .photos {
     margin: -2.5px;
     line-height: 0;
@@ -26,6 +31,7 @@
   }
   .TaxonPhoto {
     flex-grow: 1;
+    max-width: 100% !important;
   }
 }
 
@@ -81,3 +87,41 @@
   }
 }
 .photo-group.first h3 { margin-top: 0px; }
+
+.PhotoBrowser.grid {
+  .photos {
+    display: grid;
+    grid-template-columns: repeat( auto-fill, minmax( 160px, 1fr ) );
+    margin: 0;
+    gap: 5px;
+
+    @media ( max-width: $sm-max ) {
+      grid-template-columns: repeat( auto-fill, minmax( 130px, 1fr ) );
+    }
+  }
+}
+
+@media ( max-width: $sm-max ) {
+  .PhotoBrowser {
+    padding-bottom: 30px;
+  }
+
+  #controls {
+    flex-direction: column;
+
+    #layout-control {
+      margin-bottom: 10px;
+    }
+
+    #filters {
+      display: flex;
+      flex-wrap: wrap;
+      line-height: normal;
+
+      .control-group {
+        margin-bottom: 5px;
+      }
+    }
+  }
+
+}
